requirements.txt
Flask
PyPDF2
pytesseract
pdfplumber
Pillow
